api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-JWT-Extended==4.5.2  # ✅ Required for JWT authentication
Flask-Migrate==4.0.4  # ✅ Enables database migrations
